% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmonitor.R
\name{create_slot}
\alias{create_slot}
\title{Trigger slot creation on the page via the websocket}
\usage{
create_slot(name)
}
\description{
Trigger slot creation on the page via the websocket
}
\keyword{internal}
