% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmonitor.R
\name{monitor}
\alias{monitor}
\title{Monitor variables}
\usage{
monitor(..., list)
}
\arguments{
\item{...}{variables to monitor}

\item{list}{optional character vector of variable names to monitor}
}
\description{
This function starts the monitoring of a numeric variable in your environment
in the RStudio viewer pane or an external browser window. Only one connection
is supported.
}
