% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmonitor.R
\name{monitor_option}
\alias{monitor_option}
\title{Set monitor option}
\usage{
monitor_option(length = 150, xtime = FALSE, width = 450,
  height = 200, sizing = "static")
}
\arguments{
\item{length}{integer how many values to store per variable}

\item{xtime}{true/false the x-axis scale either time or history number}

\item{width}{width of each plot in pixels (min. 50)}

\item{height}{height of each plot in pixels (min. 50)}

\item{sizing}{static or dynamic. Dynamic autoscales the width based on window}
}
\description{
This function controls several options for displaying variable monitor graphs
in the currently opened browser window.
}
