% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmonitor.R
\name{mtr_stop}
\alias{mtr_stop}
\title{Stop serving the websocket and webpage for the monitor -- clean up}
\usage{
mtr_stop()
}
\description{
Stop serving the websocket and webpage for the monitor -- clean up
}
\keyword{internal}
