% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmonitor.R
\name{push_data}
\alias{push_data}
\title{Send a datapoint for a variable via the websocket}
\usage{
push_data(name, data)
}
\description{
Send a datapoint for a variable via the websocket
}
\keyword{internal}
