% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmonitor.R
\name{remove_slot}
\alias{remove_slot}
\title{Trigger slot removal on the page via the websocket}
\usage{
remove_slot(name)
}
\description{
Trigger slot removal on the page via the websocket
}
\keyword{internal}
