% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmonitor.R
\name{unmonitor}
\alias{unmonitor}
\title{Stop monitoring variables}
\usage{
unmonitor(..., list)
}
\arguments{
\item{...}{variables to monitor}

\item{list}{optional character vector of variable names to monitor}
}
\description{
This function stops the monitoring of a variable being monitored and removes
it from the active browser window or RStudio viewer pane.
}
